Revert "REFACTOR : ddl, dummy.sql 수정 및 배송 서비스 폴더 이름 수정 및 import 경로 변경"
This reverts commit 87a763939afa058f730fd5e71d874f8ed2ab55d7.
