+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/application/dto/DeliveryRouteResponseDto.java
@@ -1,7 +1,7 @@
 package com.sixlab.logistics.delivery_service.delivery.application.dto;
 
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryRoute;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryRouteStatus;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryRoute;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryRouteStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
