+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/application/dto/DeliveryStatusRequestDto.java
@@ -1,9 +1,11 @@
 package com.sixlab.logistics.delivery_service.delivery.application.dto;
 
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryStatus;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.UUID;
 
 @Data
 @AllArgsConstructor
