+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/application/dto/DeliveryStatusResponseDto.java
@@ -1,6 +1,6 @@
 package com.sixlab.logistics.delivery_service.delivery.application.dto;
 
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryStatus;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryStatus;
 import lombok.Data;
 import lombok.NoArgsConstructor;
