+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/application/service/DeliveryRouteService.java
@@ -4,9 +4,9 @@ import com.sixlab.logistics.common.shared.exception.ResourceNotFoundException;
 import com.sixlab.logistics.common.shared.security.Role;
 import com.sixlab.logistics.common.shared.security.UserDetailsImpl;
 import com.sixlab.logistics.delivery_service.delivery.application.dto.*;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.Delivery;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryRoute;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryRouteStatus;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.Delivery;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryRoute;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryRouteStatus;
 import com.sixlab.logistics.delivery_service.delivery.domain.repository.DeliveryRepository;
 import com.sixlab.logistics.delivery_service.delivery.domain.repository.DeliveryRouteRepository;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.HubService;
