+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/application/service/DeliveryService.java
@@ -4,14 +4,16 @@ import com.sixlab.logistics.common.shared.exception.ResourceNotFoundException;
 import com.sixlab.logistics.common.shared.security.Role;
 import com.sixlab.logistics.common.shared.security.UserDetailsImpl;
 import com.sixlab.logistics.delivery_service.delivery.application.dto.*;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.Delivery;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.DeliveryStatus;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.Delivery;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.DeliveryStatus;
 import com.sixlab.logistics.delivery_service.delivery.domain.repository.DeliveryRepository;
+import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.CompanyClient;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.CompanyService;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.HubService;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.dto.CompanyResponseDto;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.dto.HubManagerResponseDto;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.client.dto.HubRouteResponseDto;
+import com.sixlab.logistics.delivery_service.deliveryAgent.domain.repository.DeliveryAgentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
