+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/domain/repository/DeliveryRepository.java
@@ -2,7 +2,9 @@ package com.sixlab.logistics.delivery_service.delivery.domain.repository;
 
 import com.sixlab.logistics.delivery_service.delivery.application.dto.DeliveryResponseDto;
 import com.sixlab.logistics.delivery_service.delivery.application.dto.DeliverySearchDto;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.Delivery;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.Delivery;
+import com.sixlab.logistics.delivery_service.deliveryAgent.application.dto.DeliveryAgentResponseDto;
+import com.sixlab.logistics.delivery_service.deliveryAgent.application.dto.DeliveryAgentSearchDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
