+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/infrastructure/jpa/DeliveryJpaRepository.java
@@ -1,6 +1,10 @@
 package com.sixlab.logistics.delivery_service.delivery.infrastructure.jpa;
 
-import com.sixlab.logistics.delivery_service.delivery.domain.model.Delivery;
+import com.sixlab.logistics.delivery_service.delivery.application.dto.DeliveryResponseDto;
+import com.sixlab.logistics.delivery_service.delivery.application.dto.DeliverySearchDto;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.Delivery;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
