+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/delivery/infrastructure/repository/DeliveryRepositoryImpl.java
@@ -2,9 +2,12 @@ package com.sixlab.logistics.delivery_service.delivery.infrastructure.repository
 
 import com.sixlab.logistics.delivery_service.delivery.application.dto.DeliveryResponseDto;
 import com.sixlab.logistics.delivery_service.delivery.application.dto.DeliverySearchDto;
-import com.sixlab.logistics.delivery_service.delivery.domain.model.Delivery;
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.Delivery;
 import com.sixlab.logistics.delivery_service.delivery.domain.repository.DeliveryRepository;
 import com.sixlab.logistics.delivery_service.delivery.infrastructure.jpa.DeliveryJpaRepository;
+import com.sixlab.logistics.delivery_service.deliveryAgent.application.dto.DeliveryAgentResponseDto;
+import com.sixlab.logistics.delivery_service.deliveryAgent.application.dto.DeliveryAgentSearchDto;
+import com.sixlab.logistics.delivery_service.deliveryAgent.infrastructure.repository.DeliveryAgentQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
