+++ services/delivery-service/src/main/java/com/sixlab/logistics/delivery_service/deliveryAgent/infrastructure/repository/DeliveryAgentRepositoryImpl.java
@@ -1,5 +1,8 @@
 package com.sixlab.logistics.delivery_service.deliveryAgent.infrastructure.repository;
 
+import com.sixlab.logistics.delivery_service.delivery.domain.entity.Delivery;
+import com.sixlab.logistics.delivery_service.delivery.domain.repository.DeliveryRepository;
+import com.sixlab.logistics.delivery_service.delivery.infrastructure.jpa.DeliveryJpaRepository;
 import com.sixlab.logistics.delivery_service.deliveryAgent.application.dto.DeliveryAgentResponseDto;
 import com.sixlab.logistics.delivery_service.deliveryAgent.application.dto.DeliveryAgentSearchDto;
 import com.sixlab.logistics.delivery_service.deliveryAgent.domain.entity.DeliveryAgent;
